Save of next phase of axis changes to allow for a large number of axes by using hash tables instead of brute force axis comparisions when reading HDF5, add error handling for allocation failures on variable length arrays, fix a warning on size_t long comparison -- HJB

+++ CBFlib_bleeding_edge/src/cbf_simple.c
@@ -1539,7 +1539,7 @@ extern "C" {
                 
                 cbf_reportnez(cbf_get_longvalue(handle,&index),error);
                 
-                if (index < 1 || index > rank) error |= CBF_FORMAT;
+                if (index < 1 || index > (long)rank) error |= CBF_FORMAT;
                 
                 cbf_reportnez(cbf_find_column(handle,"size"),error);
                 
@@ -1578,8 +1578,8 @@ extern "C" {
                 
                 cbf_reportnez(cbf_get_longvalue(handle,&precedence),error);
 
-                if (index < 1 || index > rank
-                    || precedence < 1 || precedence > rank) error |= CBF_FORMAT;
+                if (index < 1 || index > (long)rank
+                    || precedence < 1 || precedence > (long)rank) error |= CBF_FORMAT;
                 
                 if (!error) {
